code review h.w.4
