--- models/contacts.json
@@ -1,1 +1,0 @@
-[{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"6","name":"ssss3s","email":"gg@gma3il.com"},{"id":"10","name":"ssses3s","email":"gg@gma3eil.com"},{"id":"11","name":"ssses3s","email":"gg@gma3eil.com"},{"id":"12","name":"ssses3s","email":"gg@gma3eil.com"},{"id":"5","name":"dad2222a1","email":"hh@gmail.com","phone":"55555555"},{"id":"13","name":"ssses3s","email":"gg@gma3eil.com"}]
